Removed Backend folder
Removed backend folder to make more sense. Instead the whole App will be here counting the UI and other stuff that it might need
