--- backend/core/src/DataStructures.h
@@ -1,17 +1,0 @@
-#pragma once
-#include <QString>
-struct Balances
-{
-	int index;
-	QString BalanceAmount;
-	QString AssetType;
-	QString AssetCode;
-	QString AssetIssuer;
-};
-
-struct StellarInfo
-{
-	QString AccID;
-	QString PublicKey;
-	QString SecretKey;
-};
