--- backend/core/src/main.cpp
@@ -1,13 +1,0 @@
-#include "SkySend_DemoVS.h"
-#include <QtWidgets/QApplication>
-
-
-int main(int argc, char *argv[])
-{
-	QApplication a(argc, argv);
-	SkySend_DemoVS w;
-
-	
-	w.show();
-	return a.exec();
-}
